model/requirements.txt
flask
flask-cors
python-dotenv
numpy
pandas
langchain
langchain-community
langchain-groq
langchain-openai
openai
